fix: f*****ng ts errors
