+++ src/core/api/index.ts
@@ -12,9 +12,6 @@ export * from './auth/auth.types';
 export * from './geo/index.api';
 export * from './geo/index.types';
 
-export * from './rounds/index.api';
-export * from './rounds/index.types';
-
 export interface PaginateReq {
   page?: number;
   limit?: number;
