+++ src/core/api/projects/index.types.ts
@@ -1,5 +1,28 @@
-import { Identity, PaginateResV3, Round } from '..';
+import { Identity, PaginateResV3 } from '..';
 import { Media } from '../media/index.types';
+
+export interface Round {
+  id: string;
+  name: string;
+  pool_amount: number;
+  cover_id: string | null;
+  cover: string | null;
+  voting_start_at: Date;
+  voting_end_at: Date;
+  submission_start_at: string;
+  submission_end_at: string;
+  total_donations: number;
+  total_votes: number;
+  total_projects: number;
+  voting_announce_at: string | null;
+  created_at: string;
+  updated_at: string;
+}
+
+export type RoundsRes = {
+  data: Round[];
+  total: number;
+};
 
 export type ProjectStatus = 'ACTIVE' | 'DRAFT';
 
@@ -116,4 +139,5 @@ export interface Donation {
     user_id: string;
   };
 }
+
 export type DonationRes = PaginateResV3<Donation>;
