+++ src/modules/Projects/containers/ProjectsList/index.types.ts
@@ -1,5 +1,3 @@
-import { Round } from 'src/core/api/rounds/index.types';
-
 export interface ProjectsListProps {
   hasTitle?: boolean;
   roundId: string;
